+++ python_modules/dagster-graphql/dagster_graphql_tests/graphql/test_auto_materialize_asset_evaluations.py
@@ -3,7 +3,7 @@ from datetime import datetime
 import dagster._check as check
 from dagster import AssetKey
 from dagster._core.definitions.asset_reconciliation_sensor import (
-    AssetReconciliationCursor,
+    AssetDaemonCursor,
     AutoMaterializeAssetEvaluation,
 )
 from dagster._core.definitions.auto_materialize_condition import (
@@ -313,7 +313,7 @@ class TestAutoMaterializeAssetEvaluations(ExecutingGraphQLContextTestMatrix):
 
     def test_current_evaluation_id(self, graphql_context: WorkspaceRequestContext):
         graphql_context.instance.daemon_cursor_storage.set_cursor_values(
-            {CURSOR_KEY: AssetReconciliationCursor.empty().serialize()}
+            {CURSOR_KEY: AssetDaemonCursor.empty().serialize()}
         )
 
         results = execute_dagster_graphql(
@@ -330,7 +330,7 @@ class TestAutoMaterializeAssetEvaluations(ExecutingGraphQLContextTestMatrix):
 
         graphql_context.instance.daemon_cursor_storage.set_cursor_values(
             {
-                CURSOR_KEY: AssetReconciliationCursor.empty()
+                CURSOR_KEY: AssetDaemonCursor.empty()
                 .with_updates(0, {}, set(), {}, 42, None, [], 0)  # type: ignore
                 .serialize()
             }
