+++ python_modules/dagster/dagster/_utils/caching_instance_queryer.py
@@ -9,6 +9,8 @@ from typing import (
     Mapping,
     Optional,
     Sequence,
+    Set,
+    Tuple,
     Union,
     cast,
 )
@@ -556,6 +558,147 @@ class CachingInstanceQueryer(DynamicPartitionsStore):
                 else None
                 for partition_key, tags in query_result.items()
             }
+
+    def _is_child_planned_for_latest_run_of_parent(
+        self,
+        child_key: AssetKey,
+        child_partition: str,
+        parent_key: AssetKey,
+        newly_updated_parent_partitions: PartitionsSubset,
+    ) -> bool:
+        """Returns a boolean representing if the given child asset partition is planned to be
+        executed in the latest run of of its parent partition for the given parent key.
+        """
+        if (
+            # if child has a different partitions def than the parent, then it must
+            # have been executed in a different run
+            self.asset_graph.get_partitions_def(child_key)
+            != self.asset_graph.get_partitions_def(parent_key)
+            # if child partition key is not the same as any newly materialized
+            # parent key, then it could not have been executed in the same run as
+            # its parent
+            or child_partition not in newly_updated_parent_partitions
+            # if child partition is not failed or in progress, then even if it was
+            # executed in the same run, we can filter it out later with an is_reconciled
+            # check (cheaper than the below logic)
+            or child_partition not in self.get_failed_or_in_progress_subset(asset_key=child_key)
+        ):
+            return False
+        else:
+            # manually query to see if this asset partition was intended to be
+            # executed in the same run as its parent
+            latest_partition_record = check.not_none(
+                self.get_latest_materialization_or_observation_record(
+                    AssetKeyPartitionKey(parent_key, child_partition)
+                )
+            )
+            if not self.is_asset_planned_for_run(
+                latest_partition_record.run_id, AssetKeyPartitionKey(child_key, child_partition)
+            ):
+                return False
+        return True
+
+    def asset_partitions_with_newly_updated_parents(
+        self,
+        after_cursor: Optional[int],
+        target_asset_keys: AbstractSet[AssetKey],
+        target_asset_keys_and_parents: AbstractSet[AssetKey],
+    ) -> Tuple[AbstractSet[AssetKeyPartitionKey], Optional[int]]:
+        """Finds asset partitions in the given selection whose parents have been materialized since
+        latest_storage_id.
+
+        Returns:
+            - A set of asset partitions.
+            - The latest observed storage_id across all relevant assets. Can be used to avoid scanning
+                the same events the next time this function is called.
+        """
+        result_asset_partitions: Set[AssetKeyPartitionKey] = set()
+        result_latest_storage_id = after_cursor
+
+        for asset_key in target_asset_keys_and_parents:
+            if self.asset_graph.is_source(asset_key) and not self.asset_graph.is_observable(
+                asset_key
+            ):
+                continue
+
+            # the set of asset partitions which have been updated since the latest storage id
+            new_asset_partitions = self.get_asset_partitions_updated_after_cursor(
+                asset_key=asset_key,
+                asset_partitions=None,
+                after_cursor=after_cursor,
+            )
+            if not new_asset_partitions:
+                continue
+
+            partitions_def = self.asset_graph.get_partitions_def(asset_key)
+            if partitions_def is None:
+                latest_record = check.not_none(
+                    self.get_latest_materialization_or_observation_record(
+                        AssetKeyPartitionKey(asset_key)
+                    )
+                )
+                for child in self.asset_graph.get_children_partitions(
+                    dynamic_partitions_store=self,
+                    current_time=self.evaluation_time,
+                    asset_key=asset_key,
+                ):
+                    if child.asset_key in target_asset_keys and not self.is_asset_planned_for_run(
+                        latest_record.run_id, child
+                    ):
+                        result_asset_partitions.add(child)
+            else:
+                partitions_subset = partitions_def.empty_subset().with_partition_keys(
+                    [
+                        asset_partition.partition_key
+                        for asset_partition in new_asset_partitions
+                        if asset_partition.partition_key is not None
+                        and partitions_def.has_partition_key(
+                            asset_partition.partition_key,
+                            dynamic_partitions_store=self,
+                            current_time=self.evaluation_time,
+                        )
+                    ]
+                )
+                for child in self.asset_graph.get_children(asset_key):
+                    child_partitions_def = self.asset_graph.get_partitions_def(child)
+                    if child not in target_asset_keys:
+                        continue
+                    elif not child_partitions_def:
+                        result_asset_partitions.add(AssetKeyPartitionKey(child, None))
+                    else:
+                        # we are mapping from the partitions of the parent asset to the partitions of
+                        # the child asset
+                        partition_mapping = self.asset_graph.get_partition_mapping(child, asset_key)
+                        child_partitions_subset = (
+                            partition_mapping.get_downstream_partitions_for_partitions(
+                                partitions_subset,
+                                downstream_partitions_def=child_partitions_def,
+                                dynamic_partitions_store=self,
+                                current_time=self.evaluation_time,
+                            )
+                        )
+                        for child_partition in child_partitions_subset.get_partition_keys():
+                            # only add this to the set if the asset is not planned to be executed
+                            # in the same run as its newly-updated parent
+                            if not self._is_child_planned_for_latest_run_of_parent(
+                                child_key=child,
+                                child_partition=child_partition,
+                                parent_key=asset_key,
+                                newly_updated_parent_partitions=partitions_subset,
+                            ):
+                                result_asset_partitions.add(
+                                    AssetKeyPartitionKey(child, child_partition)
+                                )
+
+            result_latest_storage_id = max(
+                result_latest_storage_id or 0,
+                self.get_latest_materialization_or_observation_storage_id(
+                    AssetKeyPartitionKey(asset_key)
+                )
+                or 0,
+            )
+
+        return (result_asset_partitions, result_latest_storage_id)
 
     def get_asset_partitions_updated_after_cursor(
         self,
