+++ python_modules/dagster/dagster_tests/definitions_tests/asset_reconciliation_tests/test_auto_observe.py
@@ -1,10 +1,10 @@
 from dagster import AssetKey, DagsterInstance, observable_source_asset
-from dagster._core.definitions.asset_graph import AssetGraph
-from dagster._core.definitions.asset_reconciliation_sensor import (
-    AssetReconciliationCursor,
+from dagster._core.definitions.asset_daemon_context import (
+    AssetDaemonCursor,
     get_auto_observe_run_requests,
-    reconcile,
 )
+from dagster._core.definitions.asset_graph import AssetGraph
+from dagster._core.definitions.asset_reconciliation_sensor import reconcile
 from pytest import fixture
 
 
@@ -107,7 +107,7 @@ def test_reconcile():
         asset_graph=asset_graph,
         target_asset_keys=set(),
         instance=instance,
-        cursor=AssetReconciliationCursor.empty(),
+        cursor=AssetDaemonCursor.empty(),
         materialize_run_tags=None,
         observe_run_tags={"tag1": "tag_value"},
     )
